Redo the timeout handling
The previous timeout handler did not work when we had a background activity
that was party visible, e.g. when the forground activity was a dialog. In
this case we cannot rely on the onPause()/onResume() timings, particularly
when the device was rotated. The (a) solution to this is to keep a counter
that tracks the number of onResume vs. onPause calls and only start the
timer when the number of resumes equals the number of pauses. This patch
implements this functionality.